Find the token that can reach the repo
Also, rescue `Octokit::Unauthorized` when checking if config is
reachable, because when the token was removed on GitHub's side that is
the error that will be raised (i.e. 401 - Bad credentials).

In `GithubApi` make `client` method private, since we shouldn't be
calling it directly -- defeats the purpose of our wrapper.